+++ src/constants.js
@@ -18,5 +18,5 @@ module.exports = {
   METRICS_STATUS_CANARY_ACCEPTED: 'canary-accepted',
   METRICS_STATUS_CANARY_REJECTED: 'canary-rejected',
 
-  METRICS_LABEL_DEFAULT: 'default',
+  METRICS_LABEL_DEFAULT: 'default-reject',
 };