+++ src/instrumenter.js
@@ -57,7 +57,7 @@ class Instrumenter {
     this.hitCounter = new PrometheusClient.Counter({
       name: 'divvy_hits_total',
       help: 'Counter of total HITs to Divvy.',
-      labelNames: ['status', 'type', 'rule_label'],
+      labelNames: ['status', 'rule_label'],
     });
 
     this.errorCounter = new PrometheusClient.Counter({
@@ -91,13 +91,14 @@ class Instrumenter {
   /**
    * Record a HIT operation.
    * @param {string} status The status of the hit, either "accepted" or "rejected".
-   * @param {string} type   The match type, either "rule", "default", or "none".
    * @param {string} label   The matching rule label, or null.
    */
-  countHit(status, type, label) {
+  countHit(status, label) {
     this.statsd.increment(`hit.${status}`);
-    this.statsd.increment(`hit.${status}.${type}`);
-    this.hitCounter.labels(status, type, label || '').inc();
+    if (label) {
+      this.statsd.increment(`hit.${status}.${label}`);
+    }
+    this.hitCounter.labels(status, label || '').inc();
   }
 
   /**